LayoutTests/imported/w3c/web-platform-tests/xhr/event-loadstart-upload-expected.txt


FAIL XMLHttpRequest: The send() method: Fire a progress event named loadstart on upload object (synchronous flag is unset) assert_equals: upload.onloadstart: event.total expected 7 but got 0
